refactor: split async wasm host ports
